requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
python-dotenv

# AI/ML
torch
transformers
sentence-transformers
faiss-cpu # Hoặc faiss-cpu nếu không có GPU
pandas
numpy
pyarrow
fastparquet

# Search
elasticsearch